[Enh](Firenvim): Added as lazy module.
